requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
python-dotenv==1.2.2
python-multipart==0.0.24
pydantic==2.12.5
langchain-core==1.2.28
langchain-community==0.4.1
langchain-groq==1.1.2
langchain-huggingface==1.2.1
huggingface-hub==1.8.0
duckduckgo-search==8.1.1
sentence-transformers==5.4.0
torch==2.11.0
pdf2image==1.17.0
sarvamai==0.1.27
